Fixed various backward compatibility issues caused by RDS pricing pull request. Also fixed the disk unit issue (method was for MB but result was in GB).
