+++ lib/amazon-pricing.rb
@@ -74,4 +74,179 @@ module AwsPricing
     }
 
   end
+
+
+  class Ec2PriceList < PriceList
+    
+    def initialize
+      @_regions = {}
+      get_ec2_on_demand_instance_pricing
+      get_ec2_reserved_instance_pricing
+      fetch_ec2_ebs_pricing
+    end
+
+    protected
+
+    @@OS_TYPES = [:linux, :mswin, :rhel, :sles, :mswinSQL, :mswinSQLWeb]
+    @@RES_TYPES = [:light, :medium, :heavy]
+
+    def get_ec2_on_demand_instance_pricing
+      @@OS_TYPES.each do |os|
+        fetch_ec2_instance_pricing(EC2_BASE_URL + "json/#{os}-od.json", :ondemand, os)
+      end
+    end
+
+    def get_ec2_reserved_instance_pricing
+      @@OS_TYPES.each do |os|
+        @@RES_TYPES.each do |res_type|
+          fetch_ec2_instance_pricing(EC2_BASE_URL + "json/#{os}-ri-#{res_type}.json", res_type, os)
+        end
+      end
+    end
+
+    # Retrieves the EC2 on-demand instance pricing.
+    # type_of_instance = :ondemand, :light, :medium, :heavy
+    def fetch_ec2_instance_pricing(url, type_of_instance, operating_system)
+      res = fetch_url(url)
+      res['config']['regions'].each do |reg|
+        region_name = reg['region']
+        region = find_or_create_region(region_name)
+        # e.g. type = {"type"=>"hiCPUODI", "sizes"=>[{"size"=>"med", "valueColumns"=>[{"name"=>"mswinSQL", "prices"=>{"USD"=>"N/A"}}]}, {"size"=>"xl", "valueColumns"=>[{"name"=>"mswinSQL", "prices"=>{"USD"=>"2.427"}}]}]}
+        reg['instanceTypes'].each do |type|
+          # e.g. size = {"size"=>"xl", "valueColumns"=>[{"name"=>"mswinSQL", "prices"=>{"USD"=>"2.427"}}]}
+          type['sizes'].each do |size|
+            begin
+              api_name, name = Ec2InstanceType.get_name(type["type"], size["size"], type_of_instance != :ondemand)
+              
+              region.add_or_update_ec2_instance_type(api_name, name, operating_system, type_of_instance, size)
+            rescue UnknownTypeError
+              $stderr.puts "WARNING: encountered #{$!.message}"
+            end
+          end
+        end
+      end
+    end
+
+    def fetch_ec2_ebs_pricing
+      res = fetch_url(EC2_BASE_URL + "pricing-ebs.json")
+      res["config"]["regions"].each do |ebs_types|
+        region = get_region(ebs_types["region"])
+        region.ebs_price = EbsPrice.new(region, ebs_types)
+      end
+    end
+
+  end
+
+
+
+  class RdsPriceList < PriceList
+    
+    def initialize
+      @_regions = {}
+       get_rds_on_demand_instance_pricing
+       get_rds_reserved_instance_pricing
+    end
+
+    protected
+
+    @@DB_TYPE = [:mysql, :oracle, :sqlserver]
+    @@RES_TYPES = [:light, :medium, :heavy]
+    
+    @@OD_DB_DEPLOY_TYPE = {
+                           :mysql=> {:mysql=>["standard","multiAZ"]},
+                           :oracle=> {:oracle=>["li-standard","li-multiAZ"], :oracle_byol=>["byol-standard","byol-multiAZ"]},
+                           :sqlserver=> {:sqlserver=>["li-se"], :sqlserver_express=>["li-ex"], :sqlserver_web=>["li-web"], :sqlserver_byol=>["byol"]}
+                        }
+
+
+    @@RESERVED_DB_DEPLOY_TYPE = {
+                           :oracle=> {:oracle=>"li", :oracle_byol=>"byol"},
+                           :sqlserver=> {:sqlserver=>"li-se", :sqlserver_express=>"li-ex", :sqlserver_web=>"li-web", :sqlserver_byol=>"byol"}
+                          }
+
+    
+    def is_multi_az?(type)
+      return true if type.match("multiAZ")
+      false
+    end                                  
+
+    def get_rds_on_demand_instance_pricing
+      @@DB_TYPE.each do |db|
+        @@OD_DB_DEPLOY_TYPE[db].each {|db_type, db_instances|
+          db_instances.each do |dp_type|
+            if db == :mysql or db == :oracle
+              fetch_on_demand_rds_instance_pricing(RDS_BASE_URL+"#{db}/pricing-#{dp_type}-deployments.json",:ondemand, db_type)
+            elsif db == :sqlserver
+              fetch_on_demand_rds_instance_pricing(RDS_BASE_URL+"#{db}/sqlserver-#{dp_type}-ondemand.json",:ondemand, db_type)
+            end
+          end
+        }
+      end
+    end
+
+    def get_rds_reserved_instance_pricing
+       @@DB_TYPE.each do |db|
+        if db == :mysql
+          @@RES_TYPES.each do |res_type|
+            fetch_reserved_rds_instance_pricing(RDS_BASE_URL+"#{db}/pricing-#{res_type}-utilization-reserved-instances.json", res_type, db)
+          end
+        else
+          @@RESERVED_DB_DEPLOY_TYPE[db].each {|db_type, db_instance|
+            @@RES_TYPES.each do |res_type|
+              fetch_reserved_rds_instance_pricing(RDS_BASE_URL+"#{db}/pricing-#{db_instance}-#{res_type}-utilization-reserved-instances.json", res_type, db_type) if db == :oracle
+              fetch_reserved_rds_instance_pricing(RDS_BASE_URL+"#{db}/sqlserver-#{db_instance}-#{res_type}-ri.json", res_type, db_type) if db == :sqlserver           
+            end            
+          }
+        end
+      end
+    end
+
+    def fetch_on_demand_rds_instance_pricing(url, type_of_rds_instance, db_type)
+      res = fetch_url(url)
+      res['config']['regions'].each do |reg|
+        region_name = reg['region']
+        region = find_or_create_region(region_name)
+        reg['types'].each do |type|
+          type['tiers'].each do |tier|
+            begin
+              #
+              # this is special case URL, it is oracle - multiAZ type of deployment but it doesn't have mutliAZ attributes in json.
+              if url == "http://aws.amazon.com/rds/pricing/oracle/pricing-li-multiAZ-deployments.json"
+                isMultiAz = true
+              else
+                isMultiAz = is_multi_az? type["name"]
+              end              
+              api_name, name = RdsInstanceType.get_name(type["name"], tier["name"], type_of_rds_instance != :ondemand)
+              
+              region.add_or_update_rds_instance_type(api_name, name, db_type, type_of_rds_instance, tier, isMultiAz)
+            rescue UnknownTypeError
+              $stderr.puts "WARNING: encountered #{$!.message}"
+            end
+          end
+        end
+      end
+    end
+
+    def fetch_reserved_rds_instance_pricing(url, type_of_rds_instance, db_type)
+      res = fetch_url(url)
+      res['config']['regions'].each do |reg|
+        region_name = reg['region']
+        region = find_or_create_region(region_name)
+        reg['instanceTypes'].each do |type|
+          type['tiers'].each do |tier|
+            begin
+                isMultiAz = is_multi_az? type["type"]
+                api_name, name = RdsInstanceType.get_name(type["type"], tier["size"], true)
+                
+                region.add_or_update_rds_instance_type(api_name, name, db_type, type_of_rds_instance, tier, isMultiAz)
+            rescue UnknownTypeError
+              $stderr.puts "WARNING: encountered #{$!.message}"
+            end
+          end
+        end
+      end
+    end                              
+  end
+
+
 end
