--- lib/ec2-amazon-pricing.rb
@@ -1,64 +1,0 @@
-require "amazon-pricing"
-module AwsPricing
-
-  class Ec2PriceList < PriceList
-    
-    def initialize
-      @_regions = {}
-      get_ec2_on_demand_instance_pricing
-      get_ec2_reserved_instance_pricing
-      fetch_ec2_ebs_pricing
-    end
-
-    protected
-
-    @@OS_TYPES = [:linux, :mswin, :rhel, :sles, :mswinSQL, :mswinSQLWeb]
-    @@RES_TYPES = [:light, :medium, :heavy]
-
-    def get_ec2_on_demand_instance_pricing
-      @@OS_TYPES.each do |os|
-        fetch_ec2_instance_pricing(EC2_BASE_URL + "json/#{os}-od.json", :ondemand, os)
-      end
-    end
-
-    def get_ec2_reserved_instance_pricing
-      @@OS_TYPES.each do |os|
-        @@RES_TYPES.each do |res_type|
-          fetch_ec2_instance_pricing(EC2_BASE_URL + "json/#{os}-ri-#{res_type}.json", res_type, os)
-        end
-      end
-    end
-
-    # Retrieves the EC2 on-demand instance pricing.
-    # type_of_instance = :ondemand, :light, :medium, :heavy
-    def fetch_ec2_instance_pricing(url, type_of_instance, operating_system)
-      res = fetch_url(url)
-      res['config']['regions'].each do |reg|
-        region_name = reg['region']
-        region = find_or_create_region(region_name)
-        # e.g. type = {"type"=>"hiCPUODI", "sizes"=>[{"size"=>"med", "valueColumns"=>[{"name"=>"mswinSQL", "prices"=>{"USD"=>"N/A"}}]}, {"size"=>"xl", "valueColumns"=>[{"name"=>"mswinSQL", "prices"=>{"USD"=>"2.427"}}]}]}
-        reg['instanceTypes'].each do |type|
-          # e.g. size = {"size"=>"xl", "valueColumns"=>[{"name"=>"mswinSQL", "prices"=>{"USD"=>"2.427"}}]}
-          type['sizes'].each do |size|
-            begin
-              api_name, name = Ec2InstanceType.get_name(type["type"], size["size"], type_of_instance != :ondemand)
-              
-              region.add_or_update_ec2_instance_type(api_name, name, operating_system, type_of_instance, size)
-            rescue UnknownTypeError
-              $stderr.puts "WARNING: encountered #{$!.message}"
-            end
-          end
-        end
-      end
-    end
-
-    def fetch_ec2_ebs_pricing
-      res = fetch_url(EC2_BASE_URL + "pricing-ebs.json")
-      res["config"]["regions"].each do |ebs_types|
-        region = get_region(ebs_types["region"])
-        region.ebs_price = EbsPrice.new(region, ebs_types)
-      end
-    end
-
-  end
-end
