+++ test/test-ec2-instance-types.rb
@@ -16,13 +16,13 @@ require 'test/unit'
 
 class TestEc2InstanceTypes < Test::Unit::TestCase
   def test_cc8xlarge_issue
-    pricing = AwsPricing::PriceList.new
+    pricing = AwsPricing::Ec2PriceList.new
     obj = pricing.get_instance_type('us-east', 'cc2.8xlarge')
     assert obj.api_name == 'cc2.8xlarge'
   end
 
   def test_name_lookup
-    pricing = AwsPricing::PriceList.new
+    pricing = AwsPricing::Ec2PriceList.new
     pricing.regions.each do |region|
       assert_not_nil region.name
 
@@ -43,13 +43,13 @@ class TestEc2InstanceTypes < Test::Unit::TestCase
 
   def test_available
     # Validate instance types in specific regions are available
-    pricing = AwsPricing::PriceList.new
+    pricing = AwsPricing::Ec2PriceList.new
     region = pricing.get_region('us-east')
     assert region.instance_type_available?('m1.large')
   end
 
   def test_fetch_all_breakeven_months
-    pricing = AwsPricing::PriceList.new
+    pricing = AwsPricing::Ec2PriceList.new
     pricing.regions.each do |region|
       region.ec2_instance_types.each do |instance|
         instance.operating_systems.each do |os|
@@ -64,7 +64,7 @@ class TestEc2InstanceTypes < Test::Unit::TestCase
   end
 
   def test_breakeven_month
-    pricing = AwsPricing::PriceList.new
+    pricing = AwsPricing::Ec2PriceList.new
     region = pricing.get_region('us-east')
     instance = region.get_ec2_instance_type('m1.large')
     bem = instance.get_breakeven_month(:linux, :heavy, :year1)
@@ -73,21 +73,21 @@ class TestEc2InstanceTypes < Test::Unit::TestCase
 
   def test_memory
     # Validate instance types in specific regions are available
-    pricing = AwsPricing::PriceList.new
+    pricing = AwsPricing::Ec2PriceList.new
     region = pricing.get_region('us-east')
     instance = region.get_ec2_instance_type('m1.large')
     assert instance.memory_in_mb == 7500
   end
 
   def test_non_standard_region_name
-    pricing = AwsPricing::PriceList.new
+    pricing = AwsPricing::Ec2PriceList.new
     region = pricing.get_region('eu-west-1')
     instance = region.get_ec2_instance_type('m1.large')
     assert instance.memory_in_mb == 7500
   end
 
   def test_ebs
-    pricing = AwsPricing::PriceList.new
+    pricing = AwsPricing::Ec2PriceList.new
     region = pricing.get_region('us-east')
     assert region.ebs_price.standard_per_gb == 0.10
     assert region.ebs_price.standard_per_million_io == 0.10
@@ -97,7 +97,7 @@ class TestEc2InstanceTypes < Test::Unit::TestCase
   end
 
   def test_ebs_not_null
-    pricing = AwsPricing::PriceList.new
+    pricing = AwsPricing::Ec2PriceList.new
     pricing.regions.each do |region|
       # Everyone should have standard pricing
       assert_not_nil region.ebs_price.standard_per_gb
@@ -107,7 +107,7 @@ class TestEc2InstanceTypes < Test::Unit::TestCase
   end
 
   def test_virtual_cores
-    pricing = AwsPricing::PriceList.new
+    pricing = AwsPricing::Ec2PriceList.new
     region = pricing.get_region('us-east')
     instance = region.get_ec2_instance_type('m1.large')
     assert instance.virtual_cores == 2
